+++ src/core/style/text_style.rs
@@ -36,7 +36,7 @@ pub trait EasyTextStyleExt: Sized {
   }
 
   /// See [`TextShadow`]
-  fn with_shadow(mut self, text_shadow: TextShadow) -> Self {
+  fn with_text_shadow(mut self, text_shadow: TextShadow) -> Self {
     self.easy_text_style_mut().text_shadow = text_shadow;
     self
   }