+++ src/widgets/containers/mod.rs
@@ -1,5 +1,6 @@
 pub mod button;
 pub mod horizontal_layout;
+pub mod radio_group;
 pub mod rich_text;
 pub mod slider;
 pub mod vertical_layout;